Enhanced usage documentation a bit.
As the plugin strips the part after the last dash of a classifier to interpret it as
the library type it must be specified or else the classifier will be truncated at the wrong
possition. - There is not check if the last part really is a type or part of the classifier.