Rename maven-nar-plugin to nar-maven-plugin
This is in line with issue #9, which is a follow-on to some discussion
on the Maven users list. The gist is that the maven-*-plugin
nomenclature is reserved for official Apache Maven plugins, and NAR is
not official -- it was intended/slated to be, but that process was never
completed nor is it planned to be any time soon.

This commit also updates the version number to 3.0.0-SNAPSHOT, since
this is obviously a breaking change. It seems appropriate in any case,
given the recent merger of several project forks. :-)

One remaining question is what the package prefix and groupId should
become, if they should even change at all.