Add a buildhive-settings.xml for use with the BuildHive job
We have a BuildHive job to provide continuous integration and testing here:

	https://buildhive.cloudbees.com/job/maven-nar/job/maven-nar-plugin/

Unfortunately, the cpptasks-parallel artifact is not deployed to where
BuildHive looks for it, and therefore the BuildHive job only succeeds
sporadically (whenever the cpptasks-parallel job has been built
successfully, for a limited amount of time).

BuildHive allows specifying an alternate settings.xml file for such
cases. Since we deployed cpptasks-parallel to the ImageJ Maven
repository, let's have such a file so that our BuildHive job fails only
when the code has a problem.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>