NarUtil: be prepared for log == null
In our very own unit test to determine whether the version of link.exe
could be determined properly, we initialize a Linker object with log ==
null. Since link.exe returns the exit code 1100, the code path that
tries to log the problematic command's output is executed, therefore
we need to fall back to stderr.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>