Change for break caused by NAR-109 fix
Removing the file extension results in the tree reporting with matching
results in search for existing targets even if the files are not the
same. Not sure on detail of original change and why file extensions
where ignored, partial fix is to not ignore extensions if order is not
used.

[JES: fixed whitespace and spelling]

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>