Do not ignore exit code in NarUtil#runCommand()
As discussed in https://github.com/maven-nar/nar-maven-plugin/pull/27
and https://github.com/maven-nar/nar-maven-plugin/issues/31 we cannot
simply expect all commands to be run via runCommand() to return an
exit code of 0: MSVC's linker, most notably, exits with at least the
codes 1140 and 1146 on two different systems when it is asked to show
its usage.

The suggested solution is to *expect* a failure when launching this
particular command, and to finally throw an exception upon unexpected
exit codes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>