Allow specifying a custom aol.properties file
Currently, new aol.properties values can be added by putting the file in the baseDirectory of the project. This is not very helpful, for example, for projects that would like to share aol.properties files.

This patch allows you to specify `nar.aolProperties` as a system define (via `-Dnar.aolProperties=path/to/aol.properties` on the command line) or by adding:
```
<nar.aolProperties>path/to/aol.properties<nar.aolProperties>
```

to the `<properties>` section of pom.xml.

When specifying a custom aol.properties file, if the file does not exist, then an exception is thrown.