Don’t run/compile tests when cross-compiling
When cross-compiling, it doesn’t make sense to run the tests (or even compile them) because they won’t actually execute.  We can determine if we are cross-compiling if `nar.arch` != `os.arch`.  In this situation, we can just skip the execution of the NarTestCompileMojo NarTestMojo.  The NarTestCompileMojo also has a skipNar flag which was not being used - and the same flag was added to the NarTestMojo (and used in both places).