NarUtil: fix java.home detection for Java7 on OS X
Starting with Java7, OS X is no longer inconsistent with every other OS:
it now appends "/jre" to the java.home system property as normal. So the
previous logic -- using the parent directory unless the OS is OS X -- is
no longer completely sound. A more robust strategy is to explicitly look
for the "/jre" suffix and strip it if present, regardless of the OS.