CaptureStreamHandler: track both stdout and stderr
The previous logic always threw away stdout if anything was present on
stderr. But we sometimes need the contents of stdout despite there also
being a nonempty stderr.

For backwards compatibility, this change leaves the behavior of
getOutput() alone, while adding getStdout() and getStderr() methods for
situations when more specifics are necessary.