CaptureStreamHandler: add execute static method
This method provides a way to easily execute a command while returning
the CaptureStreamHandler that was used to capture the output. This
provides a more flexible alternative to CaptureStreamHandler#run, which
returns only stderr or stdout contingent on whether stderr is nonempty.

Unfortunately, that mechanism left no way to interrogate stdout in the
case of a nonempty stderr. In contrast to run, The execute method
returns the CaptureStreamHandler itself, so that both stderr and stdout
can be freely accessed.