Add a unit test to ensure that javah finds all required dependencies
It appears that javah is no longer content without having all the
referred classes on the class path since Java 7. In other words, when a
class with a native method refers to a class outside the current class
path, javah now complains loudly and errors out.

Adapt the it0003/it0004 combo to test that javah finds all the classes
it requires.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>