Handle empty <include> in <javah> section gracefully
Before this patch, if you configuration contained something like

	<javah>
		<includes>
			<include></include>
		</includes>
	</javah>

the nar-javah goal would fail with a NullPointerException. Instead, we
handle such cases gracefully now, pretending that no include was specified
instead of an empty one.

This was reported as issue #17 by Christian Hammers. The resolution is
based on a suggestion by Greg Domjan.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>