Allow for getting NarInfo from dependency before packaging
Dependencies will be resolved as directories before they are packaged.
Therefore, we need to support extracting nar.properties from directories
in addition to .nar files.

While at it, refactor out a method to construct the path of the
nar.properties file, and let NarInfo's writeToFile() make the parent
directory when necessary.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>