Revert "Read the nar.properties from the ${buildDirectory}"
The regression introduced by this commit was that executables (which do
not want the version suffix in their file name) were built correctly
only until nar.properties were written into the buildDirectory: while
the executable does not want the version suffix, an associated .jar file
*does* want that suffix, and that fact gets recorded in the
nar.properties. As a consequence, subsequent compile runs of the
executable would append the version suffix to their file names!

This reverts commit 4e85422f8907d738f0e2ae7dedfe6e7950802d17 and adds
the classesDirectory field to NarPreparePackageMojo (which needs that
parameter).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>