Only use .def files with MSVC's linker for now
Apparently, gcc on Windows understands the -Wl,--output-def,<def-file>
parameter, but in general we should not hard-code parameters in the
Linker class which are specific to only one of the supported linkers.

A future commit should move the .def file handling into the sub
classes, preferably as a method with a path to a single .def file as
a parameter which does nothing in the Linker class, but only in the
subclasses which know how to treat .def files.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>