Allow additional options for nar-testCompile phase
Provide a second method to extend the default CompilerDef and LinkerDef
argument lists with a new 'testOptions' collection. Currently these
options are additive only, and are appended to the standard
configuration.

Test-specific compilation options are useful e.g. for instrumentation of
test code for coverage purposes (such as `-fprofile-arcs -ftest-coverage`
for gcov), or for enabling debug macros in test executions.