Resolve SNAPSHOT artifacts with multiple platform nar
Use getBaseVersion() instead of getVersion() when constructing file
paths for lib and include directories for nar dependencies in the
AbstractDependencyMojo.java. This change is same as #79 which probably
missing the AbstractDependencyMojo.java.

Without this change build will be failed if there have multiple nar in
the repositories.  For example:

hw-1.0-20140117.030545-15-x86-Windows-gpp-shared.nar
hw-1.0-20140117.093209-16-amd64-Linux-gpp-shared.nar

The Linux nar is built after the Windows nar. When building the windows
nar, maven will resolve the dependency base on the later timestamp which
will result in resolve failure. Using getBaseVersion will fix this
problem.

This fixes issue #80.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>