Stop suggesting that we're an official Apache project
This is complicated by the fact that we rely on the maven-plugins parent
project (and it would cause substantially more problems to avoid that):
The maven-site-plugin finds it funny to inherit automatically the parent
projects' site descriptors without providing for a way to, say, remove
breadcrumbs or menus.

As a really ugly workaround – a last resort because there are strong
indicators that the maven-site-plugin's developers do not acknowledge
that limitation as a shortcoming, let alone that they plan to improve
the situation – we add CSS to hide the respective breadcrumbs and menu
items inherited from the maven-plugins project.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>