Use correct file name for history.xml files
Correct a tendency of the plugin to leave temporary XML files lying around
that ended up in the nar archive. The change here rearranges the code to
delete the temporary XML rather than leave it around.

This fixes issue #106.

Johannes Schindelin was the original author of the change. IT's pass, no
wierd files.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>