Move all plugin configuration to @nnotations.
* fix pom configuration of maven-plugin-plugin to work for annotations.
* edit all the mojos and parameters.

Since the default value of defaultValue is "", I suppressed defaultValue = "".

Since the default value of a boolean is `false`, I suppressed defaultValue = "false".

In a few places I found that the doc, the @parameter defaultValue, and the
initial setting of the property disagreed. I made them agree. All the ITs pass, but
it might disturb someone's project which is accidentally depending on the
incorect setting in the @parameter (treating the doc as the definition of
correct).