NarProcessLibraries enhanced command construction
Updated the nar-process-libraries mojo to allow more flexibility when
constructing the command line to be executed by providing a mechanism to
substitute the name of the target library file, produced by this plugin, at a
point in the command other than the final argument.

To support the change above, additional configuration options have been included
and documented accordingly.  The template of configuration options is given
below.

<processLibraries>
    <libraryToken/>
    <commands>
        <command>
            <libraryType/>
            <executable/>
            <arguments>
                <argument/>
            </arguments>
        </command>
    </commands>
</processLibraries>