NarProcessLibraries enhanced command construction
Updated the nar-process-libraries mojo to allow more flexibility when
constructing the command line to be executed by substituting ${library}
in the configuration with the name of the native compile artifact.

To support the change above, additional configuration options have been
included and documented accordingly.  The template of configuration
options is given below.

<processLibraries>
    <libraryToken/>
    <commands>
        <command>
            <libraryType/>
            <executable/>
            <arguments>
                <argument/>
            </arguments>
        </command>
    </commands>
</processLibraries>