Tentative alternative to the library token
This commit is a combination of a partial revert and an alternative,
interpolating the hard-coded ${nar.library} string in the
nar-process-libraries configuration.

If accepted, it needs to be split into a documentation fix (i.e. add the
missing documentation of the configuration) and the enhancement to
interpolate nar.library specifically.

The latter commit would need to contain the information that we cannot
make use of Maven's built-in property interpolation, unfortunately,
because there are potentially multiple libraries to be processed, and we
could only set one property before the command configuration gets
interpolated.

Finally, a third commit would need to be added: an integration test. This
would be easiest as an add-on to it0010, changing the "Hello ${nar.arch}"
to "Hello ${nar.library}" and then verifying that it interpolates to the
correct path.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>