Force m2e to execute most goals (non-incrementally)
Eclipse's Maven integration requires certain extra steps of each Maven
plugin, to state how compatible the plugin is with Eclipse's version of
a lifecycle. With this change, we tell Eclipse to execute most NAR
goals, but only when building from scratch (because the steps are pretty
expensive).

Some of the goals are also triggered upon configuration, e.g. the
generation of a NarSystem class or downloading/unpacking dependencies.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>