Adding support for xlC compiler
xlC is a C/C++ compiler from IBM for AIX (an operating system from IBM). Here's a link for more info: http://www-03.ibm.com/software/products/en/xlcaix/
In aol.properties file, the xlC properties have been added.

These additions were necessary because aol.properties did not define xlC for AIX (or any other OS), it only defined g++ for AIX, although the necessary classes for xlC were defined (for example, it is found in CompilerEnum).
In Linker.java, in getVersion() method, the case of xlC was not handled.
Therefore a warning was generated saying "cannot find version number for
linker". This case is added and the warning is no longer generated.

In VisualAgeLinker.java, in the addImpliedArgs() method, the first
argument 'CCTask task' was missing. So this was not overriding the
addImpliedArgs() method of the AbstractLdLinker. This means that the
generated linking command was "xlc -shared", instead of "xlc
-qmkshrobj". Adding the argument fixed this issue.