Use the new banner in the output of `mvn site`
While at it, also clean up the links and the way the reports menu is
triggered.

Also disable the left and right banners because the parent project's
site descriptor (which maven-site-plugin automatically inherits without
providing us with a way to switch it off) would make it look as if the
NAR project is an official Apache project – which it is distinctly not,
as pointed out by Benson Margulies.

To avoid ugly "$banner.name" labels where the parent projects' site
descriptors declare left and right banners (and simply overriding via
<bannerLeft /> simply does not work), we disable the banners using CSS.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>