Allow the use of the options -p and -pg for gcc linker. Gcc linker allow the use of the options -p and -pg for debug purpose (refer to the end of the commit message for documentation).
I configured those options in a pom:
<linker>
	<name>g++</name>
	<options>
		<option>-p</option>
		<option>-pg</option>
	</options>
	...
</linker>

On linking step the options -p and -pg are preceded by "-Wl," on the
command line due to the fact that they are not registered in the
linkerOptions in the class com.github.maven_nar.cpptasks.gcc.GppLinker :
[DEBUG] g++ -Wl,-p -Wl,-pg ...

This commit add , "-p", "-pg" in linkerOptions in the classes GppLinker
and GccLinker.

-p Generate extra code to write profile information suitable for the
analysis program prof. You must use this option when compiling the
source files you want data about, and you must also use it when linking.
-pg Generate extra code to write profile information suitable for the
analysis program gprof. You must use this option when compiling the
source files you want data about, and you must also use it when linking.