Corrects the dependencies management. 1) Redefines the goals where the dependencies are downloaded and unpack, and avoid doing so several times. - Phase initialize - goal nar-download : download all the project dependencies in the local maven repository. - Phase generate-sources - goal nar-unpack : unpack in the project target folder (sub folder nar) the NAR files needed for compilation. - Phase generate-test-sources - goal nar-test-unpack : unpack in the project target folder (sub folder test-nar) the NAR files needed for tests compilations, for tests executions, and for executions.
2) Removes the execution of the goal nar-prepare-package on the phase
generate-test-sources, this was needed to get the correct library name
for tests compilation if the property output was used to rename the
library. The test compilation Mojo now gets the correct name without the
need to run nar-prepare-package.

3) On test compilation don't link the dependencies which are static
libraries as they are already included in the library for which we are
building the test executable.

4) Corrects the following error :
The SNAPSHOT libraries are stored on web repositories with the string
"SNAPSHOT" in the version number replaced by a timestamp based number.
If a project uses a SNAPSHOT library which is on a web repository and
not yet on the local repository, and defines this dependency with a
scope different than compile or test, the NAR files are unpacked in the
target folder of the project without changing back the timestamp to
"SNAPSHOT", causing compilation errors as the dependency files are
looked in "SNAPSHOT" directories.
Example :
- Library with artifactId "inetfunc" and version "1.0-SNAPSHOT".
- Stored in Sonatype Nexus repository as 
 > inetfunc-1.0-20150329.212547-2.nar
 > inetfunc-1.0-20150329.212547-2-noarch.nar
 > inetfunc-1.0-20150329.212547-2-amd64-Windows-gpp-static.nar
- Downloaded in local maven repository as
 > inetfunc-1.0-20150329.212547-2.nar
 > inetfunc-1.0-20150329.212547-2-noarch.nar
 > inetfunc-1.0-20150329.212547-2-amd64-Windows-gpp-static.nar
- Renamed in local maven repository as (good here)
 > inetfunc-1.0-SNAPSHOT.nar
 > inetfunc-1.0-SNAPSHOT-noarch.nar
 > inetfunc-1.0-SNAPSHOT-amd64-Windows-gpp-static.nar
- Unpacked in the project target folder as
 > inetfunc-1.0-20150329.212547-2-noarch
 > inetfunc-1.0-20150329.212547-2-amd64-Windows-gpp-static