Add integration tests for dependencies scopes. Create a new set of integration tests (it0028-scope) with the following content :
- In a folder "system-shared", a shared library
it0028-scope-system-shared with no dependency
(nothing special)

- In a folder "runtime-shared", a shared library
it0028-scope-runtime-shared with no dependency
(nothing special)

- In a folder "provided-static", a static library
it0028-scope-provided-static with no dependency
This library receive in parameter the file name of the library
it0028-scope-runtime-shared (the .so or .dll file) and dynamically loads
it and uses it during execution.

- In a folder "test-shared", a shared library it0028-scope-test-shared
with no dependency
(nothing special)

- In a folder "compile-shared", a shared library
it0028-scope-compile-shared with the dependencies
 > it0028-scope-system-shared with the scope system. With this scope the
property systemPath must be set, the value is
${project.basedir}/../system-shared/target/it0028-scope-system-shared-1.0-SNAPSHOT.jar
 > it0028-scope-provided-static with the scope provided
 > it0028-scope-test-shared with the scope test
 > it0028-scope-runtime-shared with the scope runtime
During compilation phase, this library is linked against
it0028-scope-system-shared and it0028-scope-provided-static.
During test compilation phase, the test executable is linked against
it0028-scope-compile-shared (which contains
it0028-scope-provided-static), it0028-scope-system-shared, and
it0028-scope-test-shared. it0028-scope-provided-static is excluded
because as it was statically linked to it0028-scope-compile-shared,
there is no use to link it again.
During test execution, the code in it0028-scope-provided-static
dynamically loads it0028-scope-runtime-shared from its library file name
(the .so or .dll file name) and invokes a method of this library. The
dependency it0028-scope-runtime-shared is needed to get the library file
(the .so or .dll file) to be unpacked in the target folder in order to
be loaded during execution, but is not needed for compilation  this
leads to the scope runtime.

- In a folder app, an application it0028-scope-app with the dependency
 > it0028-scope-compile-shared with no defined scope (the scope by
default is compile)
During compilation this executable is linked against
it0028-scope-compile-shared (which contains
it0028-scope-provided-static) and it0028-scope-system-shared as a
transitive dependency with the scope system through
it0028-scope-compile-shared. it0028-scope-provided-static and
it0028-scope-test-shared defined in it0028-scope-compile-shared are not
transitive and then are not visible from it0028-scope-app.
During execution, the code in it0028-scope-provided-static dynamically
loads it0028-scope-runtime-shared from its library file name (the .so or
.dll file name) and invokes a method of this library. The dependency
it0028-scope-runtime-shared with the scope runtime is added to
it0028-scope-app as a transitive dependency with the scope runtime
through it0028-scope-compile-shared, this allows to unpack the library
file (the .so or .dll file) in the target folder in order for it to be
loaded.