Added the new Msvc configuration class.
The configuration allows defining the MS Visual Studio version number.
This will then auto detect the Visual Studio Home. If the version is not
specified it will look for the VSxxxCOMNTOOLS environment variables to
see which
versions exist and use the maximum available version.

The configuration also allows selecting the windowsSdkVersion. If not
specified it will look in the C:\Program Files (x86)\Microsoft
SDKs\Windows or C:\Program Files (x86)\Windows Kits for available
versions.

The linker and compilers will then be configured with default includes
and library paths from both the visual studio and windows sdk homes.