Add a separate configuration class for MSVC
The configuration allows defining the MS Visual Studio version number.
This will then auto detect the Visual Studio Home. If the version is not
specified it will look for the VSxxxCOMNTOOLS environment variables to
see which versions exist and use the maximum available version.

The configuration also allows selecting the windowsSdkVersion. If not
specified it will look in the

C:\Program Files (x86)\Microsoft SDKs\Windows

or

C:\Program Files (x86)\Windows Kits

directories for available versions.

The linker and compilers will then be configured with default includes
and library paths from both the visual studio and windows sdk homes.

For example.

<msvc>
  <version>10.0</version>
  <windowsSdkVersion>7.1A</windowsSdkVersion>
</msvc>