Remove the property final on the @Parameter attributes.
In a pom I defined a library type as static, and during the execution of
the plugin the library type was set to shared which is the default
value.
By removing the keyword "final" on the attribute "type" in the class
Library, the build was back to normal. The attribute "type" has the
@Parameter annotation and thus its value must be set from the pom by
maven, but the property final seemed to prevent that.

So I removed the keyword "final" on all the @Parameter attributes.