Remove /Fa compile option as it isn't multi host friendly
Added /Fa option for completeness, but causing trouble when building
across multiple hosts.

Assuming many using Maven would be building on multiple hosts
http://stackoverflow.com/questions/36058315/building-c-static-lib-to-support-ltcg-and-linking-on-different-hosts

/Fa is somewhat useful to troubleshoot a code generation problem. 
But that's where the buck stops, its been a while since the assembly 
code could be properly assembled. 
It is strictly for diagnostic purposes and is therefore completely 
useless on a build server.

Those wishing to use /Fa should add the option through a profile.