Correction for mingw on Windows. nar-maven-plugin doesn't work anymore on Windows with mingw, because it tries to run the msvc command "link" during the initialization of the class Msvc in AbstractNarMojo.validate() (even if we don't use msvc).
Correction : in Msvc.init() I make the same test
"msvc".equalsIgnoreCase(mojo.getLinker().getName()) than in other
methods of the class (configureCCTask and configureLinker for example).
This way if the name of the linker if forced to "g++" in the pom (for
mingw), Msvc.init() doesn't do anything as if we were not on Windows and
thus doesn't quit on error.