POM: update developers and contributors
The developers and contributors sections were highly inaccurate.

Let's give credit where credit is due: Mark Donszelmann created the NAR
project, so should be listed in contributors as the founder, even though
he is not part of the current development team pledged to support it.

In general, the developers section now lists the people who are
currently active in supporting NAR, and the contributors section now
lists the people who have done substantial work (>50 commits) but are
not currently active in supporting the project.