Add 'sharedObjectName' flag to enable shared archives on AIX
This enhancement allows users to build shared archives out of their
shared library projects. By setting the sharedObjectName property users
can set the name of the shared object file which will be added to a
shared archive using the output file naming rules before being added to
nar file. This commit also adds a test and documentation for this flag
as well as two more flags which were recently added but not documented.